requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-PyMongo==2.3.0
requests==2.32.3
Pillow==10.4.0
Werkzeug==3.0.4
certifi==2024.8.30
idna==3.10
urllib3==2.2.3
charset-normalizer==3.4.0
gunicorn